+++ src/persistence/mongoose/utils/index.ts
@@ -3,9 +3,9 @@ import { Config } from 'convict';
 
 import connect from './connect';
 import queries from '../queries';
-import { DbClient } from '../../../utils/args';
+import { DbClient, Dict } from '../../../utils/args';
 
-export const schema = (client: DbClient): any => queries(client); // eslint-disable-line @typescript-eslint/no-explicit-any
+export const schema = (client: DbClient): Dict => queries(client); // eslint-disable-line @typescript-eslint/no-explicit-any
 
 export const close = (mongo: Mongoose): Promise<void> => mongo.disconnect();
